apps/backend/requirements.txt
# TestGen API — Python dependencies
#
# Install:
#   pip install -r requirements.txt
#   playwright install chromium   ← required for page inspection

# ─── Web framework ────────────────────────────────────────────────────────────
fastapi==0.115.0
uvicorn[standard]==0.32.0
python-multipart==0.0.12      # multipart/form-data for file uploads

# ─── Validation / settings ────────────────────────────────────────────────────
pydantic==2.9.2
pydantic-settings==2.5.2

# ─── HTTP client ──────────────────────────────────────────────────────────────
httpx==0.27.0

# ─── Browser automation — page inspection ─────────────────────────────────────
playwright==1.48.0
# After first install, run:  playwright install chromium

# ─── Gemini / Vertex AI ───────────────────────────────────────────────────────
# google-generativeai: API-key auth (Google AI Studio keys, AIzaSy...)
# google-cloud-aiplatform: Vertex AI auth (service account / ADC)
# Both use the same Gemini models; swap via GEMINI_BACKEND in .env
google-generativeai==0.8.3
google-cloud-aiplatform>=1.60.0

# ─── Dev / tooling ────────────────────────────────────────────────────────────
ruff==0.7.0
pytest==8.3.3
pytest-asyncio==0.24.0

# ─── Future dependencies (commented until needed) ─────────────────────────────
# google-cloud-aiplatform>=1.60.0  # Vertex AI service-account auth
# sqlalchemy[asyncio]==2.0.30
# alembic==1.13.1
# redis==5.0.4
# celery==5.4.0
